Script and document build
Adds a makefile to build a static version on linux, as just running
`cargo build` will use a dynamically linked openssl.  It uses a
container with an Rust Alpine linux image to have a static version of
the openssl libs and a musl-based system to ensure a fully static binary
is generated.