fix: ignore paths inside checkums files
Checksums files can include a path component. Mise is an example of
project publishing checksums files with a path compenent ('./' in this
case).
We parse the checksum file and look for the downloaded filename, so
there's no use of the path. So we remove it from the filename before we
work with.